Refac(v0.2): refactor transforms and clean-up (#888)
## Description

The `v0.1` transforms had several left over from past features:
- `N2VManipulate`, `XYFlip` and `XYRandomRotate90` are child of
`TransformConfig`, but it only brings a `model_dump` and is not used
anywhere else.
- `Compose` accepts `Normalize` as a transform.
- `N2VManipulate` exists in numpy and torch versions, alongside the
pixel manipulation utilities.

This PR does the following:
- Copy all deprecated transforms utilities to `careamics.compat` (e.g.
numpy n2v manipulate)
- Rename torch manipulation to the name without `_torch`
- Remove `TransformConfig` and implement the `model_dump` without
`"name"` (for class instantiation from the pydantic model) in
`N2VManipulate`, `XYFlip` and `XYRandomRotate90`
- Remove `Normalize` from the accepted transforms of `Compose` (old
`Compose` moved to `careamics.compat`)
- Removed some useless test

---------
