+++ src/careamics/compat/config/augmentations/normalize_config.py
