+++ src/careamics/compat/dataset/in_memory_tiled_pred_dataset.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 from numpy.typing import NDArray
 from torch.utils.data import Dataset
 
+from careamics.compat.config.augmentations import NormalizeConfig
 from careamics.compat.config.data import InferenceConfig
 from careamics.compat.config.data.tile_information import TileInformation
-from careamics.config.augmentations import NormalizeConfig
-from careamics.transforms import Compose
+from careamics.compat.transforms.compose import Compose
 from careamics.utils.reshape_array import reshape_array
 
 from .tiling import extract_tiles
