+++ src/careamics/compat/dataset/iterable_pred_dataset.py
@@ -9,10 +9,10 @@ from typing import Any
 from numpy.typing import NDArray
 from torch.utils.data import IterableDataset
 
+from careamics.compat.config.augmentations import NormalizeConfig
 from careamics.compat.config.data import InferenceConfig
-from careamics.config.augmentations import NormalizeConfig
+from careamics.compat.transforms.compose import Compose
 from careamics.file_io.read import read_tiff
-from careamics.transforms import Compose
 
 from .dataset_utils import iterate_over_files
