+++ src/careamics/compat/transforms/__init__.py
@@ -1,0 +1,1 @@
+"""Deprecated module."""
