+++ src/careamics/compat/transforms/normalize.py
