+++ src/careamics/config/augmentations/__init__.py
@@ -1,21 +1,15 @@
 """CAREamics transformation Pydantic models."""
 
 __all__ = [
-    "NORM_AND_SPATIAL_UNION",
     "SPATIAL_TRANSFORMS_UNION",
     "N2VManipulateConfig",
-    "NormalizeConfig",
-    "TransformConfig",
     "XYFlipConfig",
     "XYRandomRotate90Config",
 ]
 
 
 from .n2v_manipulate_config import N2VManipulateConfig
-from .normalize_config import NormalizeConfig
-from .transform_config import TransformConfig
 from .transform_unions import (
-    NORM_AND_SPATIAL_UNION,
     SPATIAL_TRANSFORMS_UNION,
 )
 from .xy_flip_config import XYFlipConfig
