+++ src/careamics/config/data/data_config.py
@@ -21,12 +21,12 @@ from pydantic import (
     model_validator,
 )
 
+from careamics.config.augmentations import XYFlipConfig, XYRandomRotate90Config
 from careamics.config.support import SupportedData
+from careamics.config.utils.random import generate_random_seed
+from careamics.config.validators import check_axes_validity, check_czi_axes_validity
 from careamics.utils import BaseEnum
 
-from ..augmentations import XYFlipConfig, XYRandomRotate90Config
-from ..utils.random import generate_random_seed
-from ..validators import check_axes_validity, check_czi_axes_validity
 from .normalization_config import NormalizationConfig
 from .patch_filter import (
     MaskFilterConfig,
