+++ src/careamics/transforms/__init__.py
@@ -1,22 +1,13 @@
 """Transforms that are used to augment the data."""
 
 __all__ = [
-    "Compose",
-    "Denormalize",
-    "ImageRestorationTTA",
     "N2VManipulate",
-    "N2VManipulateTorch",
-    "Normalize",
-    "TrainDenormalize",
+    "Transform",
     "XYFlip",
     "XYRandomRotate90",
-    "get_all_transforms",
 ]
 
-from .compose import Compose, get_all_transforms
 from .n2v_manipulate import N2VManipulate
-from .n2v_manipulate_torch import N2VManipulateTorch
-from .normalize import Denormalize, Normalize, TrainDenormalize
-from .tta import ImageRestorationTTA
+from .transform import Transform
 from .xy_flip import XYFlip
 from .xy_random_rotate90 import XYRandomRotate90
