+++ tests/compat/config/augmentations/test_n2v_manipulate_config.py
@@ -1,7 +1,7 @@
 import pytest
 
+from careamics.compat.transforms.n2v_manipulate import N2VManipulate
 from careamics.config.augmentations import N2VManipulateConfig
-from careamics.transforms import N2VManipulate
 
 
 def test_odd_roi_and_mask():
@@ -17,12 +17,6 @@ def test_odd_roi_and_mask():
 
     with pytest.raises(ValueError):
         N2VManipulateConfig(roi_size=3, struct_mask_span=6)
-
-
-def test_extra_parameters():
-    """Test that errors are thrown if we pass extra parameters."""
-    with pytest.raises(ValueError):
-        N2VManipulateConfig(roi_size=3, struct_mask_span=7, extra_param=1)
 
 
 def test_comptatibility_with_transform():
