+++ tests/compat/transforms/test_old_manipulate_n2v.py
@@ -1,6 +1,8 @@
+import numpy as np
 import pytest
 import torch
 
+from careamics.compat.transforms.n2v_manipulate import N2VManipulate as N2VManipulateNpy
 from careamics.config.augmentations import N2VManipulateConfig
 from careamics.config.support import SupportedPixelManipulation
 from careamics.transforms import N2VManipulate
@@ -8,9 +10,33 @@ from careamics.transforms import N2VManipulate
 
 @pytest.mark.parametrize(
     "strategy",
-    [SupportedPixelManipulation.UNIFORM, SupportedPixelManipulation.MEDIAN],
+    [SupportedPixelManipulation.UNIFORM.value, SupportedPixelManipulation.MEDIAN.value],
 )
 def test_manipulate_n2v(strategy):
+    """Test the N2V augmentation."""
+    # create array, adding a channel to simulate a 2D image with channel last
+    array = np.arange(16 * 16).reshape((16, 16))[np.newaxis, ...]
+
+    # create augmentation
+    aug = N2VManipulateNpy(roi_size=5, masked_pixel_percentage=5, strategy=strategy)
+
+    # apply augmentation
+    augmented = aug(array)
+    assert len(augmented) == 3  # transformed_patch, original_patch, mask
+
+    # assert that the difference between the original and transformed patch are the
+    # same pixels that are selected by the mask
+    tr_path, orig_patch, mask = augmented
+    diff_coords = np.array(np.where(tr_path != orig_patch))
+    mask_coords = np.array(np.where(mask == 1))
+    assert np.array_equal(diff_coords, mask_coords)
+
+
+@pytest.mark.parametrize(
+    "strategy",
+    [SupportedPixelManipulation.UNIFORM, SupportedPixelManipulation.MEDIAN],
+)
+def test_manipulate_n2v_torch(strategy):
     """Test the N2V augmentation."""
     # Create tensor, adding a channel to simulate a 2D image with channel first
     array = torch.arange(16 * 16).reshape(1, 16, 16).float()
@@ -34,7 +60,7 @@ def test_manipulate_n2v(strategy):
     assert torch.equal(diff_coords, mask_coords)
 
 
-def test_manipulate_n2v_with_seed():
+def test_manipulate_n2v_torch_with_seed():
     """Test that seed is properly used for reproducibility."""
     array = torch.arange(16 * 16).reshape(1, 16, 16).float()
